Changing  the test files names
