+++ .github/workflows/pr_tests.yml
@@ -159,35 +159,45 @@ jobs:
             --reruns 2 \
             --junit-xml=wrapper-results-basic.xml
 
-      - name: Run wrapper tests - send e2e part 1
+      - name: Run wrapper tests - send handle and subscription
         continue-on-error: true
         env:
           PYTHONPATH: ${{ github.workspace }}/vendor/logos-delivery-python-bindings/waku
         run: |
-          pytest tests/wrappers_tests/test_send_e2e_part1.py \
+          pytest tests/wrappers_tests/test_send_handle_and_subscription.py \
             -m "not docker_required" \
             --reruns 2 \
-            --junit-xml=wrapper-results-send-part1.xml
+            --junit-xml=wrapper-results-send-handle-and-subscription.xml
 
-      - name: Run wrapper tests - send e2e part 2
+      - name: Run wrapper tests - send relay propagation
         continue-on-error: true
         env:
           PYTHONPATH: ${{ github.workspace }}/vendor/logos-delivery-python-bindings/waku
         run: |
-          pytest tests/wrappers_tests/test_send_e2e_part2.py \
+          pytest tests/wrappers_tests/test_send_relay_propagation.py \
             -m "not docker_required" \
             --reruns 2 \
-            --junit-xml=wrapper-results-send-part2.xml
+            --junit-xml=wrapper-results-send-relay-propagation.xml
 
-      - name: Run wrapper tests - send e2e part 3
+      - name: Run wrapper tests - send lightpush and edge
         continue-on-error: true
         env:
           PYTHONPATH: ${{ github.workspace }}/vendor/logos-delivery-python-bindings/waku
         run: |
-          pytest tests/wrappers_tests/test_send_e2e_part3.py \
+          pytest tests/wrappers_tests/test_send_lightpush_and_edge.py \
             -m "not docker_required" \
             --reruns 2 \
-            --junit-xml=wrapper-results-send-part3.xml
+            --junit-xml=wrapper-results-send-lightpush-and-edge.xml
+
+      - name: Run wrapper tests - send errors and concurrency
+        continue-on-error: true
+        env:
+          PYTHONPATH: ${{ github.workspace }}/vendor/logos-delivery-python-bindings/waku
+        run: |
+          pytest tests/wrappers_tests/test_send_errors_and_concurrency.py \
+            -m "not docker_required" \
+            --reruns 2 \
+            --junit-xml=wrapper-results-send-errors-and-concurrency.xml
 
       - name: Test Report
         if: always()
