+++ tests/wrappers_tests/test_send_lightpush_and_edge.py
@@ -1,253 +1,36 @@
 import base64
 import pytest
+from src.env_vars import NODE_1
 from src.steps.common import StepsCommon
 from src.libs.common import delay, to_base64
 from src.libs.custom_logger import get_custom_logger
+from src.node.waku_node import WakuNode
 from src.node.wrappers_manager import WrapperManager
 from src.node.wrapper_helpers import (
     EventCollector,
     assert_event_invariants,
     create_message_bindings,
     get_node_multiaddr,
-    wait_for_connected,
     wait_for_propagated,
     wait_for_sent,
     wait_for_error,
 )
-from tests.wrappers_tests.conftest import build_node_config
+from src.test_data import DEFAULT_CLUSTER_ID
+from tests.wrappers_tests.conftest import build_node_config, free_port
 
 logger = get_custom_logger(__name__)
 
 PROPAGATED_TIMEOUT_S = 30.0
-SENT_TIMEOUT_S = 10.0
 NO_SENT_OBSERVATION_S = 5.0
 SENT_AFTER_STORE_TIMEOUT_S = 60.0
 OVERSIZED_PAYLOAD_BYTES = 200 * 1024
 RECOVERY_TIMEOUT_S = 45.0
 SERVICE_DOWN_SETTLE_S = 3.0
 
-# MaxTimeInCache from send_service.nim.
-MAX_TIME_IN_CACHE_S = 60.0
-# Extra slack to cover the background retry loop tick after the window expires.
-CACHE_EXPIRY_SLACK_S = 10.0
-ERROR_AFTER_CACHE_EXPIRY_TIMEOUT_S = MAX_TIME_IN_CACHE_S + CACHE_EXPIRY_SLACK_S
-RETRY_WINDOW_EXPIRED_MSG = "Unable to send within retry time window"
-
-
-class TestS02AutoSubscribeOnFirstSend(StepsCommon):
-    """
-    S02 — Auto-subscribe on first send.
-    Sender never calls subscribe_content_topic() before send().
-    The send API must auto-subscribe to the content topic used in the message.
-    Expected: send() returns Ok(RequestId), message_propagated arrives.
-    """
-
-    def test_s02_send_without_explicit_subscribe(self, node_config):
-        sender_collector = EventCollector()
-
-        node_config.update(
-            {
-                "relay": True,
-                "store": False,
-                "lightpush": False,
-                "filter": False,
-                "discv5Discovery": False,
-                "numShardsInNetwork": 1,
-            }
-        )
-
-        sender_result = WrapperManager.create_and_start(
-            config=node_config,
-            event_cb=sender_collector.event_callback,
-        )
-        assert sender_result.is_ok(), f"Failed to start sender: {sender_result.err()}"
-
-        with sender_result.ok_value as sender:
-            peer_config = {
-                **node_config,
-                "staticnodes": [get_node_multiaddr(sender)],
-                "portsShift": 1,
-            }
-
-            peer_result = WrapperManager.create_and_start(config=peer_config)
-            assert peer_result.is_ok(), f"Failed to start relay peer: {peer_result.err()}"
-
-            with peer_result.ok_value:
-                assert wait_for_connected(sender_collector) is not None, "Sender did not reach Connected/PartiallyConnected state"
-
-                message = create_message_bindings(
-                    payload=to_base64("S02 auto-subscribe test payload"),
-                    contentTopic="/test/1/s02-auto-subscribe/proto",
-                )
-
-                send_result = sender.send_message(message=message)
-                assert send_result.is_ok(), f"send() failed: {send_result.err()}"
-
-                request_id = send_result.ok_value
-                assert request_id, "send() returned an empty RequestId"
-
-                propagated = wait_for_propagated(
-                    collector=sender_collector,
-                    request_id=request_id,
-                    timeout_s=PROPAGATED_TIMEOUT_S,
-                )
-                assert propagated is not None, (
-                    f"No message_propagated event within {PROPAGATED_TIMEOUT_S}s. " f"Collected events: {sender_collector.events}"
-                )
-                assert propagated["requestId"] == request_id
-
-                error = wait_for_error(sender_collector, request_id, timeout_s=0)
-                assert error is None, f"Unexpected message_error event: {error}"
-
-
-class TestS06CoreSenderRelayOnly(StepsCommon):
-    """
-    S06 — Core sender with relay peers only, no store.
-    Sender has local relay enabled and is connected to one relay peer.
-    Expected: send() returns Ok(RequestId), message_propagated event arrives,
-    no message_sent (store disabled), no message_error.
-    """
-
-    def test_s06_relay_propagation_without_store(self, node_config):
-        sender_collector = EventCollector()
-
-        node_config.update(
-            {
-                "relay": True,
-                "store": False,
-                "lightpush": False,
-                "filter": False,
-                "discv5Discovery": False,
-                "numShardsInNetwork": 1,
-                "reliabilityEnabled": True,
-            }
-        )
-
-        sender_result = WrapperManager.create_and_start(
-            config=node_config,
-            event_cb=sender_collector.event_callback,
-        )
-        assert sender_result.is_ok(), f"Failed to start sender: {sender_result.err()}"
-
-        with sender_result.ok_value as sender:
-            peer_config = {
-                **node_config,
-                "staticnodes": [get_node_multiaddr(sender)],
-                "portsShift": 1,
-            }
-
-            peer_result = WrapperManager.create_and_start(config=peer_config)
-            assert peer_result.is_ok(), f"Failed to start relay peer: {peer_result.err()}"
-
-            with peer_result.ok_value:
-                assert wait_for_connected(sender_collector) is not None, "Sender did not reach Connected/PartiallyConnected state"
-
-                message = create_message_bindings(
-                    payload=to_base64("S06 relay-only test payload"),
-                    contentTopic="/test/1/s06-relay-only/proto",
-                )
-
-                send_result = sender.send_message(message=message)
-                assert send_result.is_ok(), f"send() failed: {send_result.err()}"
-
-                request_id = send_result.ok_value
-                assert request_id, "send() returned an empty RequestId"
-
-                propagated = wait_for_propagated(
-                    collector=sender_collector,
-                    request_id=request_id,
-                    timeout_s=PROPAGATED_TIMEOUT_S,
-                )
-                assert propagated is not None, (
-                    f"No message_propagated event within {PROPAGATED_TIMEOUT_S}s. " f"Collected events: {sender_collector.events}"
-                )
-                assert propagated["requestId"] == request_id
-
-                error = wait_for_error(sender_collector, request_id, timeout_s=0)
-                assert error is None, f"Unexpected message_error event: {error}"
-
-                sent = wait_for_sent(sender_collector, request_id, timeout_s=0)
-                assert sent is None, f"Unexpected message_sent event (store is disabled): {sent}"
-
-                assert_event_invariants(sender_collector, request_id)
-
-
-class TestS07CoreSenderRelayAndStore(StepsCommon):
-    """
-    S07 — Core sender with relay peers and store peer, reliability enabled.
-    Sender relays message to a store-capable peer; delivery service validates
-    the message reached the store via p2p reliability check.
-    Expected: Propagated, then Sent.
-    """
-
-    def test_s07_relay_propagation_with_store_validation(self, node_config):
-        sender_collector = EventCollector()
-
-        node_config.update(
-            {
-                "relay": True,
-                "store": False,
-                "lightpush": False,
-                "filter": False,
-                "discv5Discovery": False,
-                "numShardsInNetwork": 1,
-                "reliabilityEnabled": True,
-            }
-        )
-
-        sender_result = WrapperManager.create_and_start(
-            config=node_config,
-            event_cb=sender_collector.event_callback,
-        )
-        assert sender_result.is_ok(), f"Failed to start sender: {sender_result.err()}"
-
-        with sender_result.ok_value as sender:
-            peer_config = {
-                **node_config,
-                "staticnodes": [get_node_multiaddr(sender)],
-                "portsShift": 1,
-                "store": True,
-            }
-
-            peer_result = WrapperManager.create_and_start(config=peer_config)
-            assert peer_result.is_ok(), f"Failed to start store peer: {peer_result.err()}"
-
-            with peer_result.ok_value:
-                message = create_message_bindings(
-                    payload=to_base64("S07 relay+store test payload"),
-                    contentTopic="/test/1/s07-relay-store/proto",
-                )
-
-                send_result = sender.send_message(message=message)
-                assert send_result.is_ok(), f"send() failed: {send_result.err()}"
-
-                request_id = send_result.ok_value
-                assert request_id, "send() returned an empty RequestId"
-
-                propagated = wait_for_propagated(
-                    collector=sender_collector,
-                    request_id=request_id,
-                    timeout_s=PROPAGATED_TIMEOUT_S,
-                )
-                assert propagated is not None, (
-                    f"No message_propagated event within {PROPAGATED_TIMEOUT_S}s. " f"Collected events: {sender_collector.events}"
-                )
-                assert propagated["requestId"] == request_id
-
-                sent = wait_for_sent(
-                    collector=sender_collector,
-                    request_id=request_id,
-                    timeout_s=SENT_TIMEOUT_S,
-                )
-                assert sent is not None, (
-                    f"No message_sent event within {SENT_TIMEOUT_S}s after propagation. " f"Collected events: {sender_collector.events}"
-                )
-                assert sent["requestId"] == request_id
-
-                error = wait_for_error(sender_collector, request_id, timeout_s=0)
-                assert error is None, f"Unexpected message_error event: {error}"
-
-                assert_event_invariants(sender_collector, request_id)
+# Default-cluster shard-0 pubsub topic; used to subscribe the S11 docker store
+# peer so it joins the same relay mesh as the wrapper nodes (wrapper config
+# uses numShardsInNetwork=1 => shard 0).
+STORE_PEER_PUBSUB_TOPIC = f"/waku/2/rs/{DEFAULT_CLUSTER_ID}/0"
 
 
 class TestRelayToLightpushFallback(StepsCommon):
@@ -417,7 +200,7 @@ class TestS10EdgeSenderLightpushOnly(StepsCommon):
     """
 
     @pytest.mark.xfail(reason="lightpush peer discovery via staticnodes is broken, see https://github.com/logos-messaging/logos-delivery/issues/3847")
-    def test_s10_edge_lightpush_propagation(self, node_config):
+    def test_s10_edge_lightpush_propagation(self):
         sender_collector = EventCollector()
 
         common = {
@@ -488,60 +271,109 @@ class TestS10EdgeSenderLightpushOnly(StepsCommon):
                     assert_event_invariants(sender_collector, request_id)
 
 
-class TestS12IsolatedSenderNoPeers(StepsCommon):
+class TestS11EdgeSenderLightpushAndStore(StepsCommon):
     """
-    S12 — Isolated sender, no peers.
-    Sender has relay enabled but zero relay peers and zero lightpush peers.
-    Expected: send() returns Ok(RequestId), but eventually a message_error
-    event arrives (no route to propagate).
+    S11 — Edge sender with lightpush path and store validation.
+    Edge sender has no local relay; it publishes via a wrapper lightpush peer
+    and validates delivery via a docker store peer. Reliability enabled.
+    Topology:
+        [LightpushPeer] wrapper, relay=True, lightpush=True, store=False
+        [StorePeer]     docker WakuNode, relay=true, store=true,
+                        dials the lightpush peer via add_peers and subscribes
+                        to the same shard-0 pubsub topic so it joins the
+                        relay mesh and archives propagated messages.
+        [Edge]          wrapper, mode="Edge",
+                        staticnodes=[lightpush_peer],
+                        storenode=store_peer,
+                        reliabilityEnabled=True
+    Expected: send() returns Ok(RequestId), Propagated arrives, then Sent
+    (store validation succeeds), no Error.
+    Purpose: edge-mode fully validated success path against a real docker
+    store node (cross-implementation check).
     """
 
-    def test_s12_send_with_no_peers_produces_error(self, node_config):
+    def test_s11_edge_lightpush_with_store_validation(self):
         sender_collector = EventCollector()
 
-        node_config.update(
-            {
-                "relay": True,
-                "store": False,
-                "lightpush": False,
-                "filter": False,
-                "discv5Discovery": False,
-                "numShardsInNetwork": 1,
-            }
+        common = {
+            "filter": False,
+            "discv5Discovery": True,
+            "numShardsInNetwork": 1,
+        }
+
+        lightpush_config = build_node_config(
+            relay=True,
+            lightpush=True,
+            store=False,
+            **common,
         )
 
-        sender_result = WrapperManager.create_and_start(
-            config=node_config,
-            event_cb=sender_collector.event_callback,
-        )
-        assert sender_result.is_ok(), f"Failed to start sender: {sender_result.err()}"
+        lightpush_result = WrapperManager.create_and_start(config=lightpush_config)
+        assert lightpush_result.is_ok(), f"Failed to start lightpush peer: {lightpush_result.err()}"
 
-        with sender_result.ok_value as sender:
-            message = create_message_bindings(
-                payload=to_base64("S12 isolated sender payload"),
-                contentTopic="/test/1/s12-isolated/proto",
+        with lightpush_result.ok_value as lightpush_peer:
+            lightpush_multiaddr = get_node_multiaddr(lightpush_peer)
+
+            # Docker store peer — real nwaku node running as the store backend.
+            # Dial the lightpush peer via add_peers and subscribe to the same
+            # shard-0 pubsub topic so it joins the relay mesh and archives
+            # messages propagated by the lightpush peer.
+            store_peer = WakuNode(NODE_1, f"s11_store_{self.test_id}")
+            store_peer.start(relay="true", store="true")
+            self.add_node_peer(store_peer, [lightpush_multiaddr])
+            store_peer.set_relay_subscriptions([STORE_PEER_PUBSUB_TOPIC])
+            store_multiaddr = store_peer.get_multiaddr_with_id()
+
+            edge_config = build_node_config(
+                mode="Edge",
+                # assuming disc5v already happened
+                staticnodes=[lightpush_multiaddr, store_multiaddr],
+                reliabilityEnabled=True,
+                **common,
             )
 
-            send_result = sender.send_message(message=message)
-            assert send_result.is_ok(), f"send() must return Ok(RequestId) even with no peers, got: {send_result.err()}"
-
-            request_id = send_result.ok_value
-            assert request_id, "send() returned an empty RequestId"
-
-            error = wait_for_error(
-                collector=sender_collector,
-                request_id=request_id,
-                timeout_s=ERROR_AFTER_CACHE_EXPIRY_TIMEOUT_S,
+            edge_result = WrapperManager.create_and_start(
+                config=edge_config,
+                event_cb=sender_collector.event_callback,
             )
-            assert error is not None, (
-                f"No message_error event within {ERROR_AFTER_CACHE_EXPIRY_TIMEOUT_S}s "
-                f"(MaxTimeInCache={MAX_TIME_IN_CACHE_S}s + slack) for isolated sender. "
-                f"Collected events: {sender_collector.events}"
-            )
-            assert error["requestId"] == request_id
+            assert edge_result.is_ok(), f"Failed to start edge sender: {edge_result.err()}"
 
-            propagated = wait_for_propagated(sender_collector, request_id, timeout_s=0)
-            assert propagated is None, f"Unexpected message_propagated event for isolated sender: {propagated}"
+            with edge_result.ok_value as edge_sender:
+                message = create_message_bindings(
+                    payload=to_base64("S11 edge lightpush + store test payload"),
+                    contentTopic="/test/1/s11-edge-lightpush-store/proto",
+                )
+
+                send_result = edge_sender.send_message(message=message)
+                assert send_result.is_ok(), f"send() failed: {send_result.err()}"
+
+                request_id = send_result.ok_value
+                assert request_id, "send() returned an empty RequestId"
+
+                propagated = wait_for_propagated(
+                    collector=sender_collector,
+                    request_id=request_id,
+                    timeout_s=PROPAGATED_TIMEOUT_S,
+                )
+                assert propagated is not None, (
+                    f"No message_propagated event within {PROPAGATED_TIMEOUT_S}s. " f"Collected events: {sender_collector.events}"
+                )
+                assert propagated["requestId"] == request_id
+
+                sent = wait_for_sent(
+                    collector=sender_collector,
+                    request_id=request_id,
+                    timeout_s=SENT_AFTER_STORE_TIMEOUT_S,
+                )
+                assert sent is not None, (
+                    f"No message_sent event within {SENT_AFTER_STORE_TIMEOUT_S}s " f"after propagation. Collected events: {sender_collector.events}"
+                )
+                assert sent["requestId"] == request_id
+
+                error = wait_for_error(sender_collector, request_id, timeout_s=0)
+                assert error is None, f"Unexpected message_error event: {error}"
+
+                assert_event_invariants(sender_collector, request_id)
 
 
 class TestS14LightpushNonRetryableError(StepsCommon):
@@ -696,5 +528,124 @@ class TestS15LightpushRetryableErrorRecovery(StepsCommon):
 
                     error = wait_for_error(sender_collector, request_id, timeout_s=0)
                     assert error is None, f"Unexpected message_error after recovery: {error}"
+
+                    assert_event_invariants(sender_collector, request_id)
+
+
+class TestS26LightpushPeerChurn(StepsCommon):
+    """
+    S26: multiple lightpush peers, the selected one disappears,
+         an alternate remains.
+
+    Topology (3 peers + sender):
+      - peer1: relay + lightpush. The lightpush server initially selected
+        by the sender. Stopped mid-test to simulate churn.
+      - relay_peer: relay-only. Kept alive throughout the test as a
+        stable gossipsub mesh neighbour, so that after peer1 disappears
+        peer2 still has a relay path to propagate the message.
+      - peer2: relay + lightpush. The surviving lightpush server that
+        must take over once peer1 is gone.
+      - sender: edge node with peer1 and peer2 as static lightpush peers.
+    """
+
+    def test_s26_lightpush_peer_churn_alternate_remains(self, node_config):
+        sender_collector = EventCollector()
+        peer1_config = {
+            **node_config,
+            "relay": True,
+            "lightpush": True,
+            "store": False,
+            "filter": False,
+            "discv5Discovery": True,
+            "numShardsInNetwork": 1,
+            "portsShift": 1,
+            "discv5UdpPort": free_port(),
+        }
+        peer1_result = WrapperManager.create_and_start(config=peer1_config)
+        assert peer1_result.is_ok(), f"Failed to start lightpush peer1: {peer1_result.err()}"
+        peer1 = peer1_result.ok_value
+
+        relay_config = {
+            **node_config,
+            "relay": True,
+            "lightpush": False,
+            "store": False,
+            "filter": False,
+            "discv5Discovery": False,
+            "numShardsInNetwork": 1,
+            "portsShift": 4,
+        }
+
+        relay_result = WrapperManager.create_and_start(config=relay_config)
+        assert relay_result.is_ok(), f"Failed to start relay peer: {relay_result.err()}"
+
+        with relay_result.ok_value as relay_peer:
+            peer2_config = {
+                **peer1_config,
+                "staticnodes": [
+                    get_node_multiaddr(peer1),
+                    get_node_multiaddr(relay_peer),
+                ],
+                "portsShift": 2,
+                "discv5UdpPort": free_port(),
+            }
+
+            peer2_result = WrapperManager.create_and_start(config=peer2_config)
+            assert peer2_result.is_ok(), f"Failed to start lightpush peer2: {peer2_result.err()}"
+
+            with peer2_result.ok_value as peer2:
+                sender_config = {
+                    **node_config,
+                    "mode": "Edge",
+                    "relay": True,
+                    "lightpush": True,
+                    "store": False,
+                    "filter": False,
+                    "discv5Discovery": False,
+                    "numShardsInNetwork": 1,
+                    "portsShift": 3,
+                    "staticnodes": [
+                        get_node_multiaddr(peer1),
+                        get_node_multiaddr(peer2),
+                    ],
+                }
+
+                sender_result = WrapperManager.create_and_start(
+                    config=sender_config,
+                    event_cb=sender_collector.event_callback,
+                )
+                assert sender_result.is_ok(), f"Failed to start sender: {sender_result.err()}"
+
+                with sender_result.ok_value as sender_node:
+                    delay(2)
+                    stop_result = peer1.stop_and_destroy()
+                    assert stop_result.is_ok(), f"Failed to stop peer1: {stop_result.err()}"
+                    delay(2)
+
+                    message = create_message_bindings()
+                    send_result = sender_node.send_message(message=message)
+                    assert send_result.is_ok(), f"send() must return Ok(RequestId) during peer churn, got: {send_result.err()}"
+
+                    request_id = send_result.ok_value
+                    assert request_id, "send() returned an empty RequestId"
+
+                    # Expect Propagated via the surviving lightpush peer (peer2).
+                    propagated_event = wait_for_propagated(
+                        collector=sender_collector,
+                        request_id=request_id,
+                        timeout_s=PROPAGATED_TIMEOUT_S,
+                    )
+                    assert propagated_event is not None, (
+                        f"No MessagePropagatedEvent within {PROPAGATED_TIMEOUT_S}s "
+                        f"after the selected lightpush peer disappeared. "
+                        f"Collected events: {sender_collector.events}"
+                    )
+
+                    error_event = wait_for_error(
+                        collector=sender_collector,
+                        request_id=request_id,
+                        timeout_s=0,
+                    )
+                    assert error_event is None, f"Unexpected message_error event during peer churn: {error_event}"
 
                     assert_event_invariants(sender_collector, request_id)
